Create more realistic events
At least two of every series.  Using more or less real content, with some experimentation on different things to include/not, styling, etc.

Also, it seems that Jekyll likes to strop off leading dates of the form YYYY-MM-DD from filenames, and will only generate one event per day.  So we've renamed a lot of files we'd been working on.