Add support for link ids on event list items
This would allow us to define anchors to allow links to target specific events within a page.  The main use case for this is to be able to support the links we published widely for the HPC-BP webinars and SWR panels on the old Wordpress site.  That would also require a redirect from the Wordpress URL, and some extra effort to get that redirect to pass the fragment identifier.  Tracking this in #18