Improve handling of tutorial titles
Include the title-type key in the title, if it exists (used by BSSw tutorials).

For standalone events (not listed as part of a parent event) add the venue to the title if it exists (used by BSSw tutorials). For subevents, we assume that such information will be obvious from the context.