Move hpcbp events into a subdirectory
This is strictly an organizational tool.  Jekyll treats the subdirectory as a part of a collection and generates outputs in the same place as the events in the parent directory.  But now we can symlink to just the hpcbp webinars and have a section collection that's only them.  This is useful for the internal helper pages we want to be able to generate.