Cleaner event listing logic
Differentiate more between top-level events (standalone and "parent conference" events) and subevents.  This cleans up some of the logic and fixes a bug in the upcoming-events list where subevents could confuse the counting of how many events are being displayed and whether a link to the full list is needed.