Move series-teaser to a more logical location
This will work as long as we don't want to use the same include elsewhere.  Originally, we had it at the top of the main event listing page, but we later removed it because it took up too much space and distracted from the primary content of that page.  Also both the upcoming and past listings have links to "explore by series".